agents/gcp_integration/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-integration"
version = "0.1.0"
description = "Integración de Google Workspace para agentes de IA"
readme = "README.md"
requires-python = ">=3.10,<4.0"
authors = [
    {name = "Autonomos_dev", email = "cesar@autonomos.dev"}
]
dependencies = [
    "langchain-google-community>=2.0.7,<3.0.0",
    "llama-index-readers-google>=0.6.0,<0.7.0",
    "langchain-core>=0.3.0,<0.4.0",
    "scikit-learn>=1.6.1,<2.0.0",
    "numpy>=2.2.3,<3.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["gcp_integration"]
